build: Update build targets to support running host specific binaries directly

+++ blueshift.py
@@ -25,6 +25,7 @@ CONTAINER_MOUNT_DIRECTORY = os.getenv("CONTAINER_MOUNT_DIRECTORY")
 BUILD_OUTPUT_DIRECTORY = os.getenv("BUILD_OUTPUT_DIRECTORY")
 RUNTIME_OUTPUT_DIRECTORY = os.getenv("RUNTIME_OUTPUT_DIRECTORY")
 TARGET_OUTPUT_DIRECTORY = os.getenv("TARGET_OUTPUT_DIRECTORY")
+PLATFORM_TAG = os.getenv("PLATFORM_TAG")
 DEBUG_SERVER_PORT_MIN = os.getenv("DEBUG_SERVER_PORT_MIN")
 DEBUG_SERVER_PORT_MAX = os.getenv("DEBUG_SERVER_PORT_MAX")
 MASTER_PORT = os.getenv("MASTER_PORT")
@@ -165,13 +166,32 @@ def initialize_vnc_display():
     atexit.register(run_cmd, ["docker", "compose", "down", "-t", "1", VNC_CONTAINER_NAME], exit_on_failure=False, stderr=subprocess.DEVNULL)
     open_vnc_display()
 
+def get_host_os():
+    system = platform.system()
+    if system == "Linux":
+        return "linux64"
+    if system == "Windows":
+        return "win64"
+    if system == "Darwin":
+        return "osx64"
+    return system
+
+def get_target(PLATFORM_TAG=""):
+    if not PLATFORM_TAG:
+        PLATFORM_TAG = get_host_os()
+    return Path(BUILD_OUTPUT_DIRECTORY, PLATFORM_TAG, TARGET_OUTPUT_DIRECTORY)
+
+def get_remote_path(build_path):
+    with Path(build_path, ".info").open() as info:
+        remote_path, _ = info.readline().split()
+    return remote_path
+
 def run(args):
     if args.local:
-        executable = Path(".", TARGET_OUTPUT_DIRECTORY, RUNTIME_OUTPUT_DIRECTORY, args.binary)
+        executable = Path(".", get_target(), RUNTIME_OUTPUT_DIRECTORY, args.binary)
         run_cmd([executable, *args.binary_args])
     else:
-        if args.vnc:
-            initialize_vnc_display()
+        if args.vnc: initialize_vnc_display()
         if args.packet_forward:
             context = subprocess.check_output(["docker", "context", "show"], text=True).strip()
             if context == "rootless":
@@ -185,7 +205,8 @@ def run(args):
 
 def debug(args):
     debug_target_path = "relwithdebinfo" if args.release else "debug"
-    executable = Path(BUILD_OUTPUT_DIRECTORY, debug_target_path, RUNTIME_OUTPUT_DIRECTORY, args.binary)
+    build_path = Path(BUILD_OUTPUT_DIRECTORY, PLATFORM_TAG, debug_target_path)
+    executable = Path(build_path, RUNTIME_OUTPUT_DIRECTORY, args.binary)
     allow_visual = not args.terminal and args.debugger == "lldb"
     stop_on_entry = "true" if args.stop_on_entry else "false"
     # use controller name as debug name for readability if debugging client
@@ -200,6 +221,11 @@ def debug(args):
         run_cmd(["lldb-server", "platform", "--listen", f"*:{args.listen}", "--server",
                  "--min-gdbserver-port", args.min_server_port, "--max-gdbserver-port", args.max_server_port])
     elif args.local:
+        # update executable to use platform specific binary
+        build_path = Path(BUILD_OUTPUT_DIRECTORY, get_host_os(), debug_target_path)
+        executable = Path(build_path, RUNTIME_OUTPUT_DIRECTORY, args.binary)
+        remote_path = get_remote_path(build_path)
+
         if allow_visual: # debug locally with codelldb
             with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
                 s.connect(CODELLDB_ADDRESS)
@@ -209,19 +235,24 @@ def debug(args):
                     token: blueshift,
                     terminal: console,
                     stopOnEntry: {stop_on_entry},
+                    sourceMap: {{ {remote_path} : {os.getcwd()} }},
                     program: {executable},
                     args: {args.binary_args}
                 }}
                 """.encode())
-        else: # debug locally in terminal
-            args_command = "--" if args.debugger == "lldb" else "--args"
-            run_cmd([args.debugger, args_command, executable, *args.binary_args])
+        elif args.debugger == "lldb": # debug locally with lldb tui
+            run_cmd(["lldb", "-o", f"settings set target.source-map {remote_path} {cwd}",
+                             "--", executable, *args.binary_args])
+        else: # debug locally with gdb tui
+            run_cmd(["gdb", "-ex", f"set substitute-path {remote_path} {cwd}",
+                            "--args", executable, *args.binary_args])
     else: # debug on container gdbserver
-        if args.vnc:
-            initialize_vnc_display()
+        if args.vnc: initialize_vnc_display()
         initialize_host()
         DEBUG_SERVER_PORT = get_free_port()
         cwd = os.getcwd()
+        remote_path = get_remote_path(build_path)
+
         # Initialize debug server
         run_cmd(["docker", "compose", "--profile", "debug", "run", "-d", "--no-deps",
                         "-p", f"{DEBUG_SERVER_PORT}:{DEBUG_SERVER_PORT}", "--name", debug_name, "--rm",
@@ -240,7 +271,7 @@ def debug(args):
                     terminal: console,
                     request: attach,
                     stopOnEntry: {stop_on_entry},
-                    sourceMap: {{ {CONTAINER_MOUNT_DIRECTORY} : {os.getcwd()} }},
+                    sourceMap: {{ {remote_path} : {cwd} }},
                     targetCreateCommands: [target create {executable}],
                     processCreateCommands: [gdb-remote localhost:{DEBUG_SERVER_PORT}]
                 }}
@@ -248,11 +279,11 @@ def debug(args):
                 time.sleep(1) # externally extend gdb-remote connection timeout
             wait_for_container_server(DEBUG_SERVER_PORT, wait_for_exit=True)
         elif args.debugger == "lldb": # debug remotely with lldb tui
-            run_cmd(["lldb", "-o", f"settings set target.source-map {CONTAINER_MOUNT_DIRECTORY} {cwd}",
-                            "-o", f"gdb-remote localhost:{DEBUG_SERVER_PORT}", "--", executable, *args.binary_args])
+            run_cmd(["lldb", "-o", f"settings set target.source-map {remote_path} {cwd}",
+                             "-o", f"gdb-remote localhost:{DEBUG_SERVER_PORT}", "--", executable, *args.binary_args])
         else: # debug remotely with gdb tui
             run_cmd(["gdb", "-ex", f"target remote localhost:{DEBUG_SERVER_PORT}",
-                            "-ex", f"set substitute-path {CONTAINER_MOUNT_DIRECTORY} {cwd}",
+                            "-ex", f"set substitute-path {remote_path} {cwd}",
                             "--args", executable, *args.binary_args])
 
 def deploy(args):
@@ -287,7 +318,7 @@ def deploy(args):
         os.environ["DEBUG_PORT"] = str(DEBUG_SERVER_PORT)
         os.environ["DEBUG_SERVER_PORT_MAX"] = str(int(DEBUG_SERVER_PORT_MIN) + num_clients + 1) # +1 for master
 
-        debug_target_path = Path(BUILD_OUTPUT_DIRECTORY, os.getenv("BUILD_TYPE").lower(), RUNTIME_OUTPUT_DIRECTORY)
+        debug_target_path = Path(BUILD_OUTPUT_DIRECTORY, PLATFORM_TAG, os.getenv("BUILD_TYPE").lower(), RUNTIME_OUTPUT_DIRECTORY)
         master_binary = Path(debug_target_path, "master")
         client_binary = Path(debug_target_path, "client")
         # start watchers to attach to each target process when spawned in container
@@ -298,25 +329,29 @@ def deploy(args):
     run_cmd(command)
 
 def build(args):
+    if args.image_build:
+        run_cmd(["docker", "compose", "build"])
+        return
+
     ARTIFACT_TYPE = args.build_type.lower()
-    TARGET_PLATFORM = args.target if args.target != "local" else ""
-    ARTIFACT_DIR = Path(BUILD_OUTPUT_DIRECTORY, TARGET_PLATFORM, ARTIFACT_TYPE)
+    PLATFORM_TAG = args.target
+    ARTIFACT_DIR = Path(BUILD_OUTPUT_DIRECTORY, PLATFORM_TAG, ARTIFACT_TYPE)
     COMPILE_DB_PATH = Path(BUILD_OUTPUT_DIRECTORY, "compile_commands.json")
     remote_args = [f"--build-type", args.build_type,
                   "--compiler", args.compiler,
                   "--parallel", str(args.parallel),
                   "--target", args.target]
-    if args.clean:
-        remote_args.append("--clean")
-        rmtree(ARTIFACT_DIR, ignore_errors=True)
+    if args.clean: remote_args.append("--clean")
 
     if args.local:
+        if args.clean: rmtree(ARTIFACT_DIR, ignore_errors=True)
+
         cpp_compiler = "-DCMAKE_CXX_COMPILER="
         c_compiler = "-DCMAKE_C_COMPILER="
         linker = "-DCMAKE_LINKER_TYPE="
         archiver = "-DCMAKE_AR="
         build_type = f"-DCMAKE_BUILD_TYPE={args.build_type}"
-        toolchain_file = "-DCMAKE_TOOLCHAIN_FILE=" + (str(Path(os.getcwd(), ".cmake", f"{args.compiler}-{TARGET_PLATFORM}.cmake")) if TARGET_PLATFORM else "")
+        toolchain_file = "-DCMAKE_TOOLCHAIN_FILE="
         if args.compiler == 'clang':
             c_compiler += "clang"
             cpp_compiler += "clang++"
@@ -327,9 +362,16 @@ def build(args):
             cpp_compiler += "g++"
             linker += "BFD"
             archiver += "ar"
+        if PLATFORM_TAG != "linux64": # only cross compile when not compiling for linux
+            toolchain_file += str(Path(os.getcwd(), ".cmake", f"{args.compiler}-{PLATFORM_TAG}.cmake"))
+        
         cmake_args = [c_compiler, cpp_compiler, linker, build_type, toolchain_file, "-Wno-dev"]
         run_cmd(["cmake", *cmake_args, "-S", ".", "-B", ARTIFACT_DIR])
-        symlink(COMPILE_DB_PATH, Path(".", TARGET_PLATFORM, ARTIFACT_TYPE, "compile_commands.json"))
+        symlink(COMPILE_DB_PATH, Path(".", PLATFORM_TAG, ARTIFACT_TYPE, "compile_commands.json"))
+        # used for clangd and debugger source maps
+        with Path(ARTIFACT_DIR, ".info").open("w") as build_info:
+            build_info.write(f"{os.getcwd()} {PLATFORM_TAG}")
+        symlink(Path(BUILD_OUTPUT_DIRECTORY, ".info"), Path(".", PLATFORM_TAG, ARTIFACT_TYPE, ".info"))
 
         # 1 <= j <= n-1 (keep 1 core free for background tasks)
         core_count = str(max(1, min(args.parallel, NUM_CORES - 1)))
@@ -337,21 +379,13 @@ def build(args):
         build_args = list(filter(None, build_args))
         run_cmd(["cmake", "--build", ".", *build_args, "-t", *args.make],
                        cwd=f"./{ARTIFACT_DIR}")
-        if not TARGET_PLATFORM:
-            # only update binary targets if built for current machine's arch
-            symlink(TARGET_OUTPUT_DIRECTORY, Path(".", ARTIFACT_TYPE), True)
+        symlink(get_target(PLATFORM_TAG), Path(".", ARTIFACT_TYPE), True)
     else: # build binaries in remote container
-        @run_as_src_user
-        def write_build_info():
-            with Path(BUILD_OUTPUT_DIRECTORY, ".info").open("w") as build_info:
-                build_info.write(f"{CONTAINER_MOUNT_DIRECTORY} {os.getenv("PLATFORM_TAG")}")
-        atexit.register(write_build_info)
         initialize_host()
-        if TARGET_PLATFORM:
-            run_cmd(["docker", "compose", "build", "builder"])
-            os.environ["PLATFORM_TAG"] = TARGET_PLATFORM
-        if args.image_build:
-            run_cmd(["docker", "compose", "build"])
+        # ensure builder base image (linux64) is built
+        run_cmd(["docker", "compose", "build", "builder"])
+        # use correct platform for building
+        os.environ["PLATFORM_TAG"] = PLATFORM_TAG
         run_cmd(["docker", "compose", "run", "--rm", "builder",
                         "build", "-l", *remote_args, *args.make])
 
@@ -359,12 +393,13 @@ def test(args):
     os.environ["GTEST_COLOR"] = "1"
     ctest_args = [args.verbose, "-R", args.tests_regex]
     ctest_args = list(filter(None, ctest_args))
-    if args.local:
+    if args.local: # TODO: fix local test execution in non container environments (map container src dir to host src dir in ctest?)
         fail_output = ["--output-on-failure"] if not args.no_output_on_failure else []
-        if not Path(TARGET_OUTPUT_DIRECTORY).exists():
+        TARGET_DIR = get_target()
+        if not TARGET_DIR.exists():
             print("No tests to run.")
             sys.exit(0)
-        run_cmd(["ctest", *fail_output, *ctest_args], cwd=f"./{TARGET_OUTPUT_DIRECTORY}")
+        run_cmd(["ctest", *fail_output, *ctest_args], cwd=f"./{TARGET_DIR}")
     else: # test in remote container
         initialize_host()
         run_cmd(["docker", "compose", "run", "--rm", "builder", "test", "-l", *ctest_args])
@@ -495,7 +530,7 @@ build_parser.add_argument("make",
                           nargs="*",
                           default=["all"])
 build_parser.add_argument("-i", "--image-build",
-                          help="build container images",
+                          help="build container images instead of source",
                           action="store_true")
 build_parser.add_argument("-c", "--compiler",
                           help="set compiler used for building",
@@ -512,8 +547,8 @@ build_parser.add_argument("-b", "--build-type",
                           default="Debug")
 build_parser.add_argument("-t", "--target",
                           help="set target platform",
-                          choices=["local", "rpi64", "win64", "osx64"],
-                          default="local")
+                          choices=["linux64", "rpi64", "win64", "osx64"],
+                          default="linux64")
 build_parser.add_argument("-v", "--verbose",
                           help="get verbose output from cmake",
                           action="store_const",
